[add/#1] hilt 기초 세팅 추가
