+++ app/src/main/java/com/smashing/app/presentation/main/MainActivity.kt
@@ -1,0 +1,25 @@
+package com.smashing.app.presentation.main
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import com.smashing.app.core.designsystem.theme.SmashingAndroidTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            SmashingAndroidTheme {
+                val appState = rememberMainAppState()
+
+                MainScreen(
+                    appState = appState,
+                )
+            }
+        }
+    }
+}
